Force-framework: renamed exception.
